Rename package locale_profile to localprofile
